+++ Assets/Resources/Sounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c089c6d8651c30941bf074561f6b5f52
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
