+++ Assets/Scripts/Attacks/AttackHoming.cs
@@ -7,6 +7,7 @@ public class AttackHoming : Attack
     {
         GameObject newMissile = Instantiate(missile, shootPos.position, Quaternion.identity) as GameObject;
 
+            EnemyAI.laserSound.Play();
         float x  = playerPos.position.x - shootPos.position.x;
         float y = playerPos.position.y - shootPos.position.y;
         if (Mathf.Abs(x) >= Mathf.Abs(y))
@@ -23,6 +24,7 @@ public class AttackHoming : Attack
         newMissile.GetComponent<Rigidbody2D>().velocity = new Vector2(x * shootSpeed * Time.fixedDeltaTime, y * shootSpeed * Time.fixedDeltaTime);
         newMissile.transform.SetParent(GameObject.Find(Constants.CANVAS_OBJECT).transform, true);
         newMissile.transform.SetSiblingIndex(4);
+        EnemyAI.laserSound.Play();
     }
 
     public static IEnumerator ShootInWaves(GameObject enemy, GameObject missile, float delay, int ammo, float shootSpeed, Transform shootPos, Transform playerPos)
