+++ Assets/Scripts/Attacks/AttackSpiral.cs
@@ -16,6 +16,7 @@ public class AttackSpiral : Attack
                 newMissile.transform.SetParent(GameObject.Find(Constants.CANVAS_OBJECT).transform, true);
                 newMissile.transform.SetSiblingIndex(4);
                 angle += 15;
+                EnemyAI.laserSound.Play();
                 yield return new WaitForSeconds(0.01f);
             }
             angle = 0;
