--- Assets/Scripts/GameScreenOverlays/BackToMenu.cs
@@ -1,20 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class BackToMenu : MonoBehaviour
-{
-    public void Back()
-    {
-        SceneManager.LoadScene(Constants.MENU_SCENE);
-    }
-
-    public void PauseGame()
-    {
-        Time.timeScale = 0;
-    }
-
-    public void ResumeGame()
-    {
-        Time.timeScale = 1;
-    }
-}
