--- Assets/Scripts/GameScreenOverlays/BackToMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f75b9f458efb978409b2328f2d617d4f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
